Move setting up of test visual studio locator to a fixture
